adds DomainScalar in new files, improves SDM, DDM
